[refactor] ChartFrame commom으로 이동
- 재사용 가능한 ChartFrame component를 common 으로 이동
