+++ src/components/common/ChartFrame.tsx
